+++ app/Services/Image/ImageHashingService.php
@@ -6,6 +6,7 @@ namespace Kami\Cocktail\Services\Image;
 
 use Thumbhash\Thumbhash;
 use Intervention\Image\Interfaces\ImageInterface;
+
 use function Thumbhash\extract_size_and_pixels_with_imagick;
 
 class ImageHashingService