+++ CHANGELOG.md
@@ -43,6 +43,10 @@ This new major release includes improvements to API specification, new minimal P
 - Added `quantity` to shopping list ingredients
 - Images are now converted to WebP format before saving to disk
 - Added `/images` endpoint, this endpoint is used to list all user uploaded images
+- Bar search tokens are now saved in database
+
+## Changes
+- Meilisearch client API keys are now generated via artisan command
 
 # v3.19.0
 ## New